feat: validations and restructure of files
