--- src/components/Auth/index.ts
@@ -1,3 +1,0 @@
-export * from '../Auth/FormLogin';
-export * from '../Auth/FormRegister';
-export * from '../Auth/ImageContent';
