+++ src/pages/auth/register.tsx
@@ -1,7 +1,7 @@
-import { FormRegister, ImageContent } from '@/components/Auth';
 import { AuthLayout } from '@/layouts/AuthLayout';
 import client from '@/public/client-register.svg';
 import { Logo } from '@/ui/Logo';
+import { FormRegister, ImageContent } from '@/views/auth/components';
 import { Box } from '@mui/system';
 import type { NextPage } from 'next';
