+++ src/views/auth/components/index.ts
@@ -1,0 +1,3 @@
+export * from './FormLogin';
+export * from './FormRegister';
+export * from './ImageContent';
